requirements.txt
requests
pytest
allure-pytest
pydantic
faker
